implement file-scoped aliases
aliases declared within a .csv or .xlsx file now apply to aliases used
inside that file.

aliases are declared in an #aliases "blueprint" -- that is, beneath an
"#aliases" modeline. any line before the next modeline that contains a
valid alias will be added to the file-scoped alias list.

aliases can be in either of two formats, and users can mix formats
freely within the same #aliases section. the first format is the same as
what is used in the aliases.txt files:

aliasname: aliasdefinition

and aliases in this format must appear in the first column of a row.

The second format has the alias name in the first column and the alias
definition in the second column, with no colon separator:

aliasname,aliasdefinition

There can be multiple #aliases sections defined in a .csv file or .xlsx
spreadsheet. The aliases are simple combined into one list. If an alias
is defined multiple times, the last definition wins.