Add mass-remove.lua, a gui script for mass removal
This gui script can be used to mark buildings and
constructions for removal in bulk, instead of one
by one. It can also be used to suspend or
unsuspend planned constructions/buildings in bulk.
Buildings/constructions that are marked for
removal can be unmarked, but planned constructions
that are removed with this tool cannot be
recovered.

The tool also provides the following visual cues
to aid in mass removal:

* Constructions that are marked for removal appear
flashing with a red 'n' symbol.
* Buildings that are marked for removal appear
flashing with a red 'x' symbol.
* Planned constructions/buildings that are
suspended appear with a flashing red 's' symbol.