replaced RETIRED_ADVENTURER flag with ADVENTURER
RETIRED_ADVENTURER is no longer set when adventurers are retired in the current version (presumably as of 0.47.01 in relation to the tactical party update). It is now solely the ADVENTURER flag that determines whether or not a nemesis is included in the unretire list (as opposed to 0.44.12, where this was due to RETIRED_ADVENTURER).