accept string data params for one-cell blueprints
before, the data param passed to the quickfort API had to contain a
coordinate map. a common case for using the API, though, is to apply a
one-cell blueprint. we can simplify the syntax for such usage by
allowing data to be just a simple string, and interpreting it as
{[0]={[0]={[0]=string}}}