add gui/workorder-details.lua: adjusts input item, material, traits (#358)
* add gui/workorder-finetune.lua

Adjust input items, material, or traits for work orders. Actual jobs created for it will inherit the details.

* use reqscript 'gui/workshop-job'

This allows to reuse previously duplicated code from `gui/workshop-job.lua`.

Also, minor wording changes and comments.

* minor changes

* make everything local

* clarify comment regarding `gui/workshop-job`

* add test

* remove tab

* update a comment

* rename to `workorder-details`, add a changelog entry

* remove trailing spaces

* use the `workorder` script to create the work order for the test

* use `gui.materials.ItemTraitsDialog`

* check temporary order is removed

* handle `confirm` plugin in test

* add a test for removing every possible trait

* use default `ItemTraitsDialog.on_select` handler

This requires DFHack/dfhack#2211

* fix merge-error in changelog.txt

* in tests, pause `confirm` plugin instead of using `delay(5)`

When the `SELECT` key is sent to the `confirm` prompt, the intercepted key isn't passed on to the underlying screen until the next rendered frame. Since we escape out of that screen before the next rendered frame, it is never sent -- until the screen is reopened. If you run the test, and it fails, and then you manually open the manager orders screen, you can see the test order there before it quickly disappears. If you pause the confirm dialog and resend `MANAGER_REMOVE`, it works.

Co-authored-by: Myk <myk002@yahoo.com>

Co-authored-by: Myk <myk002@yahoo.com>