devel/visualize-structure: dereference primitive double-pointers
For instance, passing `screen.parent` to this script is equivalent to
`screen:_field('parent')`, which effectively yields a `viewscreen**`. This is
represented in Lua as an object with a single `value` field containing the
actual viewscreen pointer. However, there is no way to dereference this pointer
via `utils.df_expr_to_ref()`. That behavior is fine for non-pointer primitives,
since there is nothing to dereference, so this change introduces a special case
for pointer primitives only.