Fix "in chest" detection, and clarify 'construction' and 'in_building'
Original code looked for `item.flags.in_chest` which I guess used to indicate an item in a chest. I cannot find any items in my save that have this set, nor any other scripts that reference it (except deep-embark, I believe erroneously) so it seems likely it is no longer used in v50.

Instead, items stored in a chest (box), bookcase, cabinet, workshop, etc. have a `building_holder` in general_refs.  This updated code checks for that to indicate if the item is in a container, and if so what type as read from `df.building_type`.

Additionally, I clarified the output for `item.flags.in_building`, which shows when an item is used to build a building (rather than being contained in the building), and likewise for `item.flags.construction`.