Minor fix
Add an argument for level to the `out` function the script provides if none is provided by the caller

I assume this default definition of the out function was in place for module purposes. As far as I can see, no scripts are using troubleshoot-item as a module at present.

It did occur to me that with the new info added this could make a good basis for an overlay script to display all this info for all items - although that may be a duplicate of `view-item-info` (which still needs updating for v50)