Add ``g`` shortcut to locate a target or selected value on the map
When a fort is loaded jump to the first of:
- the coord if the current target or the selected value is a coord
- the `pos` of the current target
- the `centerx`, `centery`, `z` of the current target
- the `pos` of the selected value
- the `centerx`, `centery`, `z` of the selected value

The last two allow browsing lists like world.units.active or
world.buildings.all and pressing g to jump to each item in the list
without opening them.
Items in containers/being carried won't be located currently - I can
add a (recursive) holder/container ref lookup if there is interest.