make launcher autocomplete case insensitive
this will exclude entries that have legitimate capital letters in them,
but the only such victim right now is RemoteFortressReader, which is a
dev tool and is hidden by default, so I'm not that worred about it. We
intend for all tools to be lower case, so this shound't cause problems
in the future. (Hi future person looking at this, trying to debug the
problem it causes)